requirements.txt
Flask
Scrapy
pymongo
pyyaml
Unidecode